requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3
scipy==1.12.0
openai>=1.50.0
python-dotenv==1.0.0
pydantic==2.5.3
Pillow==10.2.0
dtaidistance==2.3.11
